Egen klasse for saldoutregning av flerbarnsdager
